Changed the names of add two ints service and modified CMakeLists to reflect the new name
