+++ src/edit_talker_client.cpp
@@ -1,19 +1,19 @@
 #include "ros/ros.h"
-#include "beginner_tutorials/AddTwoInts.h"
+#include "beginner_tutorials/EditTalker.h"
 #include <cstdlib>
 
 int main(int argc, char **argv)
 {
-  ros::init(argc, argv, "add_two_ints_client");
+  ros::init(argc, argv, "edit_talker");
   if (argc != 2)
   {
-    ROS_INFO("usage: add_two_ints_client X");
+    ROS_INFO("usage: edit_talker_client X");
     return 1;
   }
 
   ros::NodeHandle n;
-  ros::ServiceClient client = n.serviceClient<beginner_tutorials::AddTwoInts>("add_two_ints");
-  beginner_tutorials::AddTwoInts srv;
+  ros::ServiceClient client = n.serviceClient<beginner_tutorials::EditTalker>("edit_talker");
+  beginner_tutorials::EditTalker srv;
   srv.request.a = argv[1];
   if (client.call(srv))
   {
@@ -22,7 +22,7 @@ int main(int argc, char **argv)
   }
   else
   {
-    ROS_ERROR("Failed to call service add_two_ints");
+    ROS_ERROR("Failed to call service edit_talker");
     return 1;
   }
